requirements.txt
ffmpeg-python
transformers
numpy
tqdm
pandas
nltk